requirements/testing.txt
-r base.txt

pytest==8.1.1
pytest-django==4.8.0
